test(dashboard): extract insight-filter cases to non-excluded suite
Migrates the two insight-filter assertions from
tests/relay/dashboard-edge-cases.test.ts (gated by RUN_KNOWN_BROKEN=1
due to 3 pre-existing totalSignals failures) into a new file
tests/relay/dashboard-open-findings-insight-filter.test.ts that runs in
CI default. The parent suite remains in KNOWN_BROKEN_SUITES; its
totalSignals failures are tracked separately.
